Module Lambda__Bool

exception NotABoolTerm

raised whenever trying to convert a non boolean term to boolean

val ltrue : Lambda.L.term

true lambda term

val lfalse : Lambda.L.term

false lambda term

val to_bool : Lambda.L.term -> bool

to_bool t converts a lambda term t to boolean

val of_bool : bool -> Lambda.L.term

of_bool b converts a boolean t to a lambda term

val is_bool : Lambda.L.term -> bool

is_bool t returns true if t is a boolean lambda term